Fix NPE when customizing context (#206)
If you do not provide a MavenSystemHome when creating a Runtime, attempts to customize the context on that runtime will fail with an NPE because the customize method attempts to invoke `derive` without a null guard. Aside from customizing the context, MIMA works as expected with a null maven system home, so the simplest fix is to add a null guard.

I also included a test case.