Expose resolver version
When possible. When embedded in Maven, classloader isolation
prevents us to discover, but Maven version implicitly reveals
resolver version as well (check release notes of it or check the
`/lib`). When standalone, discovery is usable and is used, as
MIMA never delivers 'full maven' so it can combine those two
with minor differences between real Maven distro releases.